backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.20
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
alembic==1.13.0
pymongo==4.6.0
motor==3.3.2

boto3~=1.34.0
# Machine Learning - PyTorch (nhẹ hơn TensorFlow)
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2

# Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pydantic==2.5.0

underthesea==6.7.0
pyvi==0.1.1

mlflow==2.9.0

beautifulsoup4==4.12.2